COLLECTING's COLLECT local function now returns the passed in argument.
The previous behavior was to return the newly created last cons of the list being collected, which is of dubious value and even potentially dangerous because this cons will be destructively modified on subsequent collections. I believe the intended semantics was that COLLECT returns an undefined value, but it seems intuitive and potentially helpful to simply return the collected value, which is safe to use.